python/DDjango/requirements.txt
appdirs==1.4.3
Django==1.11.1
django-markdown2==0.3.0
docopt==0.6.2
git-fame==1.2.0
mysqlclient==1.3.10
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
